+++ client/package.json
@@ -9,7 +9,8 @@
     "axios": "^0.27.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "^2.1.3",
+    "react-scripts": "5.0.1",
+    "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^2.1.3",
     "web-vitals": "^2.1.4"
   },
