+++ client/src/App.css
@@ -5,11 +5,3 @@ body{
   background-color: #193d85;
   font-family: "Open Sans", sans-serif;
 }
-
-.header{
-  text-align: center;
-}
-
-.row{
-  padding-bottom: 50px;
-}
