+++ client/src/App.js
@@ -1,14 +1,14 @@
-import React from "react"
-import "./App.css"
+import React from "react";
+import ToDoList from "./ToDoList"
 import {Container} from "semantic-ui-react"
-import ToDoList from "./To-Do-List"
+import "./App.css"
 
 export default function App(){
-    return(
-        <div>
-            <Container>
-                <ToDoList/>
-            </Container>
-        </div>
-    )
+  return(
+    <div>
+      <Container>
+        <ToDoList/>
+      </Container>
+    </div>
+  )
 }
