+++ client/src/ToDoList.css
@@ -1,0 +1,7 @@
+.header{
+    text-align: center;
+}
+  
+.row{
+    padding-bottom: 50px;
+}
